# openapi/babycharts-v1.yaml
openapi: 3.1.0
info:
  title: BabyCharts API
  description: Privacy-first self-hosted child growth, milestone and health tracking REST API.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Current API v1 endpoint
paths:
...
  /auth/register:
    post:
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
                  maxLength: 128
                familyName:
                  type: string
                inviteCode:
                  type: string
                setupToken:
                  type: string
      responses:
        '200':
          description: Registration successful
        '400':
          description: Validation error
        '403':
          description: Registration forbidden
  /auth/login:
    post:
      summary: Authenticate user credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                totpCode:
                  type: string
      responses:
        '200':
          description: Authentication successful or 2FA required
        '401':
          description: Invalid credentials
...
  /auth/forgot-password:
    post:
      summary: Request password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Password reset email queued
  /auth/reset-password:
    post:
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password successfully reset
...
  /media/upload:
    post:
      summary: Upload encrypted media file
      security:
        - BearerAuth: []
        - SessionCookie: []
      responses:
        '200':
          description: Media uploaded
  /media/{id}:
    get:
      summary: Download and decrypt media file
      security:
        - BearerAuth: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decrypted media file stream
    delete:
      summary: Delete media file
      security:
        - BearerAuth: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Media deleted
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionCookie:
      type: apiKey
      in: cookie
      name: babycharts_session